# use bits.Mul64 for avoiding compiler optimization