```
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/int128
               │   .old.txt   │              .new.txt               │
               │    sec/op    │    sec/op     vs base               │
```